redefine server db ListItems logic
it is now using a callback-iterator approach,
which keeps the code a lot simpler,
and which prevents the data race which was detected previously

fixes #477
